feat: Combine two different stack service implementations
