--- cli/pkg/stack_mgr/COVERAGE
@@ -1,1 +1,0 @@
-54.3
